[STORY-3.3] refactor: detekt quality improvements - class decomposition
Quality enhancements:
- Split TenLayerJwtValidator into focused validator classes (TooManyFunctions resolved)
- Created JwtFormatValidator for Layers 1-3 (format, signature, algorithm)
- Created JwtClaimsValidator for Layers 4-6 (claims, timing, issuer)
- Created JwtSecurityValidator for Layers 7-10 (revocation, roles, user, injection)
- Reduced detekt violations from 37 to 14 (remaining are security design decisions)
- Improved separation of concerns and maintainability
- Fixed parameter naming conventions (tenant_id -> tenantId)

Test status:
- Unit tests: 6 passed - core validation logic functional after refactoring
- Class structure: Better organized with single responsibility principle
- Compilation: All modules successful
- Quality: Significant improvement in code organization
