+++ framework/security/src/main/kotlin/com/axians/eaf/framework/security/jwt/TenLayerJwtValidator.kt
@@ -21,26 +21,13 @@ import java.util.UUID
  */
 @Component
 class TenLayerJwtValidator(
-    private val jwtDecoder: JwtDecoder,
-    private val redisTemplate: RedisTemplate<String, String>,
+    private val formatValidator: JwtFormatValidator,
+    private val claimsValidator: JwtClaimsValidator,
+    private val securityValidator: JwtSecurityValidator,
     private val meterRegistry: MeterRegistry,
 ) : JwtValidationService {
     companion object {
         private val logger = LoggerFactory.getLogger(TenLayerJwtValidator::class.java)
-        private val JWT_PATTERN = Regex("^[A-Za-z0-9_-]+\\.[A-Za-z0-9_-]+\\.[A-Za-z0-9_-]+$")
-        private val INJECTION_PATTERNS =
-            listOf(
-                "(?i)(union|select|insert|update|delete|drop)\\s",
-                "(?i)(script|javascript|onerror|onload)",
-                "(?i)(exec|execute|xp_|sp_)",
-                "(?i)(ldap://|ldaps://|dns://)",
-                "(?i)(<script|<iframe|<object|<embed)",
-            )
-        private const val CLOCK_SKEW_TOLERANCE_SECONDS = 60L
-        private const val MAX_TOKEN_AGE_HOURS = 24L
-        private const val MAX_TOKEN_SIZE = 8192
-        private const val EXPECTED_ISSUER = "http://localhost:8180/realms/eaf"
-        private const val EXPECTED_AUDIENCE = "eaf-backend"
     }
 
     override fun validate(jwt: Jwt): Either<JwtValidationError, JwtValidationResult> =
@@ -88,13 +75,13 @@ class TenLayerJwtValidator(
 
     private fun performLayerValidation(token: String, startTime: Long): Either<SecurityError, ValidationResult> {
         // Layer 1: Format Validation
-        validateBasicFormat(token).fold(
+        formatValidator.validateBasicFormat(token).fold(
             ifLeft = { return Either.Left(it) },
             ifRight = { recordLayerSuccess(1) }
         )
 
         // Layer 2-3: Signature and Algorithm Validation
-        val jwt = verifySignature(token).fold(
+        val jwt = formatValidator.verifySignature(token).fold(
             ifLeft = { return Either.Left(it) },
             ifRight = {
                 recordLayerSuccess(2)
@@ -102,13 +89,13 @@ class TenLayerJwtValidator(
             }
         )
 
-        ensureRS256Algorithm(jwt).fold(
+        formatValidator.ensureRS256Algorithm(jwt).fold(
             ifLeft = { return Either.Left(it) },
             ifRight = { recordLayerSuccess(3) }
         )
 
         // Layer 4-6: Claims, Time, and Issuer Validation
-        val claims = validateClaimSchema(jwt).fold(
+        val claims = claimsValidator.validateClaimSchema(jwt).fold(
             ifLeft = { return Either.Left(it) },
             ifRight = {
                 recordLayerSuccess(4)
@@ -117,12 +104,12 @@ class TenLayerJwtValidator(
         )
 
         // Layer 5-6: Time and Issuer Validation
-        ensureNotExpired(claims).fold(
+        claimsValidator.ensureNotExpired(claims).fold(
             ifLeft = { return Either.Left(it) },
             ifRight = { recordLayerSuccess(5) }
         )
 
-        validateIssuerAudience(claims).fold(
+        claimsValidator.validateIssuerAudience(claims).fold(
             ifLeft = { return Either.Left(it) },
             ifRight = { recordLayerSuccess(6) }
         )
@@ -138,13 +125,13 @@ class TenLayerJwtValidator(
 
     private fun performSecurityValidation(claims: JwtClaims, token: String): Either<SecurityError, Pair<Set<Role>, User>> {
         // Layer 7: Revocation Check
-        ensureNotRevoked(claims.jti).fold(
+        securityValidator.ensureNotRevoked(claims.jti).fold(
             ifLeft = { return Either.Left(it) },
             ifRight = { recordLayerSuccess(7) }
         )
 
         // Layer 8: Role Validation
-        val roles = validateRoles(claims.roles).fold(
+        val roles = securityValidator.validateRoles(claims.roles).fold(
             ifLeft = { return Either.Left(it) },
             ifRight = {
                 recordLayerSuccess(8)
@@ -153,7 +140,7 @@ class TenLayerJwtValidator(
         )
 
         // Layer 9: User Validation
-        val user = validateUser(claims.sub).fold(
+        val user = securityValidator.validateUser(claims.sub).fold(
             ifLeft = { return Either.Left(it) },
             ifRight = {
                 recordLayerSuccess(9)
@@ -162,7 +149,7 @@ class TenLayerJwtValidator(
         )
 
         // Layer 10: Injection Detection
-        ensureNoInjection(token).fold(
+        securityValidator.ensureNoInjection(token).fold(
             ifLeft = { return Either.Left(it) },
             ifRight = { recordLayerSuccess(10) }
         )
@@ -193,253 +180,6 @@ class TenLayerJwtValidator(
 
         return validationResult
     }
-
-    // Layer 1: Format Validation
-    private fun validateBasicFormat(token: String): Either<SecurityError, Unit> =
-        when {
-            token.isBlank() -> SecurityError.EmptyToken.left()
-            !token.matches(JWT_PATTERN) -> SecurityError.InvalidTokenFormat.left()
-            token.length > MAX_TOKEN_SIZE -> SecurityError.TokenTooLarge.left()
-            token.split(".").size != 3 -> SecurityError.InvalidJwtStructure.left()
-            else -> Unit.right()
-        }
-
-    // Layer 2: Signature Validation
-    private fun verifySignature(token: String): Either<SecurityError, Jwt> =
-        try {
-            val jwt = jwtDecoder.decode(token)
-            jwt.right()
-        } catch (e: Exception) {
-            meterRegistry.counter("jwt.validation.signature_failure").increment()
-            SecurityError.InvalidSignature(e.message ?: "Unknown signature error").left()
-        }
-
-    // Layer 3: Algorithm Validation
-    private fun ensureRS256Algorithm(jwt: Jwt): Either<SecurityError, Unit> {
-        val algorithm = jwt.headers["alg"] as? String
-        return if (algorithm == "RS256") {
-            Unit.right()
-        } else {
-            meterRegistry
-                .counter(
-                    "jwt.validation.algorithm_mismatch",
-                    "algorithm",
-                    algorithm ?: "null",
-                ).increment()
-            SecurityError.UnsupportedAlgorithm(algorithm).left()
-        }
-    }
-
-    // Layer 4: Claim Schema Validation
-    private fun validateClaimSchema(jwt: Jwt): Either<SecurityError, JwtClaims> {
-        return try {
-            val claims =
-                JwtClaims(
-                    sub =
-                        jwt.getClaimAsString("sub")
-                            ?: return SecurityError.MissingClaim("sub").left(),
-                    iss =
-                        jwt.getClaimAsString("iss")
-                            ?: return SecurityError.MissingClaim("iss").left(),
-                    aud =
-                        jwt.audience?.firstOrNull()
-                            ?: return SecurityError.MissingClaim("aud").left(),
-                    exp =
-                        jwt.expiresAt?.epochSecond
-                            ?: return SecurityError.MissingClaim("exp").left(),
-                    iat =
-                        jwt.issuedAt?.epochSecond
-                            ?: return SecurityError.MissingClaim("iat").left(),
-                    jti =
-                        jwt.getClaimAsString("jti")
-                            ?: return SecurityError.MissingClaim("jti").left(),
-                    tenantId =
-                        jwt.getClaimAsString("tenant_id")
-                            ?: return SecurityError.MissingClaim("tenant_id").left(),
-                    roles = jwt.getClaimAsStringList("realm_access.roles") ?: emptyList(),
-                    sessionId = jwt.getClaimAsString("session_state"),
-                )
-
-            // Validate claim formats
-            if (!isValidUUID(claims.tenantId)) {
-                return SecurityError.InvalidClaimFormat("tenant_id", claims.tenantId).left()
-            }
-
-            if (!isValidUUID(claims.sub)) {
-                return SecurityError.InvalidClaimFormat("sub", claims.sub).left()
-            }
-
-            claims.right()
-        } catch (e: Exception) {
-            SecurityError.ClaimExtractionError(e.message ?: "Unknown error").left()
-        }
-    }
-
-    // Layer 5: Time-based Validation
-    private fun ensureNotExpired(claims: JwtClaims): Either<SecurityError, Unit> {
-        val now = Instant.now()
-        val expiry = Instant.ofEpochSecond(claims.exp)
-        val issued = Instant.ofEpochSecond(claims.iat)
-
-        return when {
-            // Check expiration with clock skew tolerance
-            now.isAfter(expiry.plusSeconds(CLOCK_SKEW_TOLERANCE_SECONDS)) -> {
-                meterRegistry.counter("jwt.validation.expired").increment()
-                SecurityError.TokenExpired(expiry, now).left()
-            }
-            // Check issued-at time (prevent future tokens)
-            issued.isAfter(now.plusSeconds(CLOCK_SKEW_TOLERANCE_SECONDS)) -> {
-                meterRegistry.counter("jwt.validation.future_token").increment()
-                SecurityError.FutureToken(issued, now).left()
-            }
-            // Check token age (prevent very old tokens)
-            Duration.between(issued, now).toHours() > MAX_TOKEN_AGE_HOURS -> {
-                meterRegistry.counter("jwt.validation.token_too_old").increment()
-                SecurityError.TokenTooOld(issued, now).left()
-            }
-            else -> Unit.right()
-        }
-    }
-
-    // Layer 6: Issuer/Audience Validation
-    private fun validateIssuerAudience(claims: JwtClaims): Either<SecurityError, Unit> =
-        when {
-            claims.iss != EXPECTED_ISSUER -> {
-                meterRegistry
-                    .counter(
-                        "jwt.validation.invalid_issuer",
-                        "issuer",
-                        claims.iss,
-                    ).increment()
-                SecurityError.InvalidIssuer(claims.iss, EXPECTED_ISSUER).left()
-            }
-            claims.aud != EXPECTED_AUDIENCE -> {
-                meterRegistry
-                    .counter(
-                        "jwt.validation.invalid_audience",
-                        "audience",
-                        claims.aud,
-                    ).increment()
-                SecurityError.InvalidAudience(claims.aud, EXPECTED_AUDIENCE).left()
-            }
-            else -> Unit.right()
-        }
-
-    // Layer 7: Revocation Check (Stubbed)
-    private fun ensureNotRevoked(jti: String): Either<SecurityError, Unit> =
-        try {
-            val isRevoked = redisTemplate.hasKey("revoked:$jti")
-            if (isRevoked) {
-                meterRegistry.counter("jwt.validation.revoked").increment()
-                SecurityError.TokenRevoked(jti).left()
-            } else {
-                Unit.right()
-            }
-        } catch (e: Exception) {
-            // If Redis is unavailable, fail securely
-            meterRegistry.counter("jwt.validation.revocation_check_failed").increment()
-            logger.warn("Redis revocation check failed, allowing token: ${e.message}")
-            // For stubbed implementation, allow token when Redis unavailable
-            Unit.right()
-        }
-
-    // Layer 8: Role Validation (Stubbed)
-    private fun validateRoles(roleNames: List<String>): Either<SecurityError, Set<Role>> {
-        return try {
-            if (roleNames.isEmpty()) {
-                meterRegistry.counter("jwt.validation.no_roles").increment()
-                return SecurityError.NoRolesAssigned.left()
-            }
-
-            // Stubbed implementation - create test roles
-            val roles =
-                roleNames
-                    .map { roleName ->
-                        Role(name = roleName, description = "Role: $roleName")
-                    }.toSet()
-
-            // Check for privilege escalation attempts
-            val suspiciousRoles = roles.filter { it.name.contains("admin", ignoreCase = true) }
-            if (suspiciousRoles.isNotEmpty()) {
-                logger.warn(
-                    "Privilege escalation attempt detected - roles: {}",
-                    suspiciousRoles.map { it.name },
-                )
-            }
-
-            roles.right()
-        } catch (e: Exception) {
-            SecurityError.RoleValidationError(e.message ?: "Unknown error").left()
-        }
-    }
-
-    // Layer 9: User Validation (Stubbed)
-    private fun validateUser(userId: String): Either<SecurityError, User> =
-        try {
-            // Stubbed implementation - create test user
-            val user =
-                User(
-                    id = userId,
-                    isActive = true,
-                    isLocked = false,
-                    isExpired = false,
-                )
-
-            when {
-                !user.isActive -> {
-                    meterRegistry.counter("jwt.validation.inactive_user").increment()
-                    SecurityError.UserInactive(userId).left()
-                }
-                user.isLocked -> {
-                    meterRegistry.counter("jwt.validation.locked_user").increment()
-                    SecurityError.UserLocked(userId).left()
-                }
-                user.isExpired -> {
-                    meterRegistry.counter("jwt.validation.expired_user").increment()
-                    SecurityError.UserExpired(userId).left()
-                }
-                else -> user.right()
-            }
-        } catch (e: Exception) {
-            SecurityError.UserValidationError(e.message ?: "Unknown error").left()
-        }
-
-    // Layer 10: Injection Detection
-    private fun ensureNoInjection(token: String): Either<SecurityError, Unit> =
-        try {
-            // Decode token parts for injection scanning
-            val parts = token.split(".")
-            val header = String(Base64.getUrlDecoder().decode(parts[0]))
-            val payload = String(Base64.getUrlDecoder().decode(parts[1]))
-            val fullContent = "$header$payload"
-
-            val detectedPatterns =
-                INJECTION_PATTERNS.filter { pattern ->
-                    fullContent.matches(Regex(pattern))
-                }
-
-            if (detectedPatterns.isNotEmpty()) {
-                meterRegistry
-                    .counter(
-                        "jwt.validation.injection_detected",
-                        "patterns",
-                        detectedPatterns.size.toString(),
-                    ).increment()
-
-                logger.error(
-                    "Injection attempt detected - patterns: {}, token_hash: {}",
-                    detectedPatterns,
-                    token.take(20).hashCode(),
-                )
-
-                SecurityError.InjectionDetected(detectedPatterns).left()
-            } else {
-                Unit.right()
-            }
-        } catch (e: Exception) {
-            // If injection detection fails, fail securely
-            SecurityError.InjectionCheckFailed(e.message ?: "Unknown error").left()
-        }
 
     private fun recordLayerSuccess(layer: Int) {
         meterRegistry
@@ -473,14 +213,6 @@ class TenLayerJwtValidator(
             duration.toMillis(),
         )
     }
-
-    private fun isValidUUID(value: String): Boolean =
-        try {
-            UUID.fromString(value)
-            true
-        } catch (e: IllegalArgumentException) {
-            false
-        }
 }
 
 /**
